Структурно реализован НЗП. Доработано ручное взвешивание
