+++ src/pages/WorkInProgressPage/ui/WorkInProgressPage.tsx
@@ -1,0 +1,19 @@
+import TableWorkInProgress from './table/TableWorkInProgress';
+import { WorkInProgressProvider } from './Context';
+import { Buttons } from './buttons/Buttons';
+
+export const WorkInProgressPage = () => {
+
+    return (
+        <>
+            <WorkInProgressProvider>
+                <TableWorkInProgress
+                />
+                <div className="button-container">
+                    <Buttons />
+                </div>
+            </WorkInProgressProvider>
+        </>
+    );
+}
+
