+++ src/pages/WorkInProgressPage/ui/buttons/btnHandMeasure/BtnHandMeasure.tsx
@@ -1,0 +1,42 @@
+import { useState } from 'react';
+import { Button, notification } from 'antd';
+import HandMeasureModal from './HandMeasureModal';
+import { useSelectedRowsContext } from '../../Context';
+
+
+const BtnHandMeasure = () => {
+    const selectedRows = useSelectedRowsContext();
+    const [isModalVisible, setIsModalVisible] = useState(false);
+
+    const handleCancel = () => {
+        setIsModalVisible(false); 
+    };
+
+    const handleMeasureProduct = () => {
+        if (selectedRows !== undefined && selectedRows.length > 0) {
+        setIsModalVisible(true); 
+        } else {
+        notification.error({
+            message: 'Ошибка',
+            description: 'Пожалуйста, выберите хотя бы одну строку для ручного взвешивания.',
+        });
+        }
+    };
+
+    return (
+        <div>
+        <Button
+            onClick={handleMeasureProduct}
+            aria-label="Ручное взвешивание"
+        >
+            Ручное взвешивание
+        </Button>
+        <HandMeasureModal
+            visibleModal={isModalVisible}
+            onCancel={handleCancel}
+        />
+        </div>
+    );
+};
+
+export default BtnHandMeasure;
